feat(wallets): add watchAsset functionality to wallet API (#3388)
### TL;DR
This pull request introduces the `watchAsset` functionality to the injected wallets in the thirdweb package.

### What changed?
A new method `watchAsset` is added to the `createAccount` function and `Account` type. This method allows users to add an ERC-20 asset to their wallet's watchlist.

### How to test?
1. Create an account using the `createAccount` function.
2. Call the `watchAsset` method with appropriate parameters to watch an ERC-20 asset.
3. Verify if the asset is added to the wallet's watchlist.

### Why make this change?
This change is made to enhance the functionality of the injected wallets by allowing users to easily add ERC-20 assets to their watchlist, improving the user experience and making asset management more convenient.

---

<!-- start pr-codex -->

---

## PR-Codex overview
The focus of this PR is to add `watchAsset` support in injected wallet accounts.

### Detailed summary
- Added `watchAsset` method in injected wallet accounts
- Defined `WatchAssetParams` type for asset watching
- Provided example usage for `watchAsset` method

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->