fix: always use CustomThemeProvider's theme prop in React (#3416)
### Main issue

StyledComponents come with their own `theme` prop, overriding our own `theme` prop. So instead of trying to use the `theme` prop in our Button component, always use the `CustomThemeProvider` one.

### TL;DR
This PR introduces a change to ensure that the `CustomThemeProvider` and related components respect an undefined `theme` prop. This is achieved by removing the default fallback to the `dark` theme when the `theme` prop is not provided.

### What changed?
- Updated `CustomThemeProvider` to accept `undefined` as a valid `theme` value.
- Modified components that use `CustomThemeProvider` to pass the `theme` prop directly without defaulting to `dark`.
- Replaced direct import and usage of `parseTheme` in `TransactionButton` with wrapping the button in `CustomThemeProvider`.
- Removed unnecessary imports and streamlined theme handling in the `Button` component to utilize `useCustomTheme`.
- Updated `dev:dashboard` script to also filter `./packages/thirdweb`.

### How to test?
- Verify the application operates correctly when the `theme` prop is explicitly set to `undefined` or not provided.
- Ensure all components render with the default theme correctly without any regressions.

### Why make this change?
This change ensures more flexible theme handling in the application and reduces the assumption that a `dark` theme should always be the fallback. This is particularly useful for dynamic theming use cases where the theme might not always be determinable upfront.

---