fix(dashboard): file uploads work across all browsers (#3431)
### TL;DR

The FileInput component has been updated to prevent event propagation when clicked, ensuring the click event does not trigger unintended parent component behavior, and to ensure the dropzone is always present -- fixing a bug with older versions of chrome and other chromium browsers.

### What changed?

- Added `onClick={(e) => e.stopPropagation()}` to the root stack holding the file input
- Moved the input element to the root stack and removed duplicated props `getRootProps` and `getInputProps`

---

<!-- start pr-codex -->

---

## PR-Codex overview
The focus of this PR is to refactor the `FileInput` component in the dashboard app for improved structure and functionality.

### Detailed summary
- Updated `FileInput` component structure for better readability
- Added `Stack` component for layout
- Enhanced event handling for better user experience

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->