fix(mint): add ability to override transaction properties in mintTo functions (#3543)
This pull request introduces a minor update to the thirdweb library, specifically adding the feature to override transaction properties in the `mintTo` functions for ERC1155, ERC20, and ERC721 tokens. Additionally, it refines the ENS avatar resolution tests by removing an offchain record test that relies on a secret key, improving overall test reliability.

---