added cache removal when faucet transfer fails (#3726)
## Problem solved

Users rate limited when faucet fails to transfer funds e.g. faucet empty

## Changes made

- [ ] Added the mechanism to remove users from the cache when the funding transfer fails (e.g. when the faucet is empty)

## How to test

- [ ] empty the faucet or make the transfer request fail

## Contributor NFT

Paste in your wallet address below and we will airdrop you a special NFT when your pull request is merged.

```Address: ```

<!-- start pr-codex -->

---

## PR-Codex overview
The focus of this PR is to add a new function `cacheDeleteKey` to delete a cache key and update its usage in the testnet faucet route.

### Detailed summary
- Added `cacheDeleteKey` function to delete a cache key in `lib/redis`
- Updated the testnet faucet route to use `cacheDeleteKey` for cache management

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->