changed comeback tomorrow to give funds (#3744)
## Problem solved

Short description of the bug fixed or feature added

## Changes made

- [ ] come back tomorrow gives funds now

## How to test

- [ ] Automated tests: link to unit test file
- [ ] Manual tests: step by step instructions on how to test

## Contributor NFT

Paste in your wallet address below and we will airdrop you a special NFT when your pull request is merged.

```Address: ```

<!-- start pr-codex -->

---

## PR-Codex overview
This PR updates the SpinWheel component in the dashboard app to adjust the amount displayed based on the result and improve the success message.

### Detailed summary
- Updated the amount displayed based on the result in `SpinWheel.tsx`
- Improved success message with a more informative and positive message

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->