fix(connect-ui): update Connect UI tabs styling with new design enhancements (#3746)
![Screenshot 2024-07-18 at 1.44.13 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/G5kV6fAq4H1wu10SVop6/82422ccc-d794-4562-a52f-d0d282b9789b.png)

### TL;DR

Update the styling of Connect UI tabs.

### What changed?

Modified the styles within `Tabs.tsx`, including adding border-radius, setting specific colors and shadows for the selected tab button, and removing the `Line` component.

### How to test?

Verify the updated styling of tabs within the Connect UI. Ensure that the selected tab has distinct styling according to the changes made.

### Why make this change?

To improve the visual appearance and user experience of the Connect UI tabs.

---

<!-- start pr-codex -->

---

## PR-Codex overview
The focus of this PR is to update the styling of Connect UI tabs and improve the user experience in the Transactions screen.

### Detailed summary
- Updated Connect UI tabs styling for better visibility and interaction
- Renamed "Purchases" tab to "Transactions" for clarity
- Adjusted height and appearance of tabs for improved layout
- Removed unnecessary line component in Tabs
- Imported `radius` from design system for consistent styling

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->