fix(dashboard): audited contract icon alignment (#3793)
### TL;DR

Improve the UI layout of the contract-card component by adjusting spacing and removing unnecessary Skeleton wrappers.

### What changed?

The `gap` property for the `Flex` component was increased from 0 to 1 to provide better spacing. The `Skeleton` wrappers around the `Icon` and `Text` components were removed for a cleaner design.

### How to test?

1. Go to the dashboard and navigate to the explore section.
2. Check the contract-card components and observe the new spacing and the absence of loading skeletons around the `Icon` and `Text`.

### Why make this change?

The change is aimed at improving the user interface by making the contract-card component cleaner and visually more appealing.

---