Update middleware.ts (#3798)
## Problem solved

Short description of the bug fixed or feature added

## Changes made

- [ ] Public API changes: list the public API changes made if any
- [ ] Internal API changes: explain the internal logic changes

## How to test

- [ ] Automated tests: link to unit test file
- [ ] Manual tests: step by step instructions on how to test

## Contributor NFT

Paste in your wallet address below and we will airdrop you a special NFT when your pull request is merged.

```Address: ```

<!-- start pr-codex -->

---

## PR-Codex overview
The focus of this PR is to improve the validation of `possibleChainId` in the middleware by checking if it is an integer using `Number.isInteger()` instead of `Number.parseInt()`.

### Detailed summary
- Changed `Number.parseInt()` to `Number()` for `possibleChainId` assignment.
- Updated validation to use `Number.isInteger()` for `possibleChainId`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->