[Playground] Fix: No longer auto disconnect from modal (#4513)
Users are reporting the inability to login on the modal as a "bug" and it's causing confusion while trying to reproduce actual bugs. This PR allows connecting on the playground modal as is the expected behavior.

<!-- start pr-codex -->

---

## PR-Codex overview
The focus of this PR is to update the `navLinks.ts` file and refactor the `RightSection.tsx` component in the `playground-web` app.

### Detailed summary
- Updated the name from "in-App wallet" to "In-App wallet" in `navLinks.ts`
- Refactored the `RightSection` component in `RightSection.tsx`
- Removed the `useDisconnect` hook and refactored the disconnect logic in `RightSection.tsx`
- Improved conditional rendering for the modal in `RightSection.tsx`

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->